AGX Opcode Table
These are values that *may* work within M2 architecture. Since this is a matter of bruteforcing our way into bare metal, we point to a shitton of registers in hopes that we are able to send memory to these locations.

I've refused to yet test these in functions provided by Alyssa since I don't want to begin breaking things. However, next build of this should give me some ideas of if these attempts work

Because Apple has these registers a missing there is a lot of fundamental things we need to do to try and reverse engineer things without trying to decompile the entirety of Apple's RTK and Mac OS X codebase.

Of course this would be so much easier if they cared to release documentation, because sadly, it's a mystery that involves a good amount of detective work