Commands for GPU (Untested)
This will give us a good idea if we can even talk with hardware by sending it a bunch of data and seeing what it can do.

The plan is to have these commands ran in my input section, so we can test things and then provided additional debugging parameters if needed. But I will be sticking to Alyssa's codebase as much as possible when it comes to overall logic.