Debugging Update #01 - Register Debugger
This allows us to view what is under the hood with exactly the information we send to the register, and it gives us an idea of exactly where the problem lies:

With this:
Testing one instruction
# AA 81 50 A0 0A 00
+fadd.32 w0// Raw bytes: AA 81 50 A0 0A 00
// src0_packed: 0x050A
// src0 decoded: type=1, reg=16, size32=1, abs=0, neg=1, unk=0
, wunk1:16.neg // Value is UNKNOWN. s.type == 1
// src1_packed: 0x000A
// src1 decoded: type=0, reg=0, size32=1, abs=0, neg=1, unk=0
, w#0.neg // s.type = 0: this is an immediate value

Whatever we are shoving into the register gets returned as a type 1, meaning that our value is completely unknown. Something is happening as I attempt to move the data into this register.

This could be where we are off in register???????? This is where my knowledge becomes shit and trying to figure out a solution becomes even more difficult, because I don't know if I'm just sending the incorrect data, or that I'm at the incorrect register