perfmon.c (Needs main removal)
So outside of perfmon having a main that conflicts with the project main (because we cannot have two main functions), this is what we will use in our menu system of our main.c file. We essentially will dedicate a menu to the implementation of all these functions that allow us to look further under the hood.