perfmon_private.h
Ah yes, the private file we are trying to reverse engineer. If you are unaware, Apple is a very open source company with their entire OSs, at least to the extent that they should be concerning security. There are plenty of private files that you can access, Apple will just expose only the amount you can see, and then place the code wrapped in an invisibility cloak for all the security vulnerable parts.

While I do not think that this is close to the original file, this is at least a reversed engineered version that handles all the perfmon definitions so we get zero errors in VSCode.